Retry on failure downloading binary from github.com
We've recently seen instances where the `curl` command times out
attempting to download the test-reporter binary from github.com:

> curl: (7) Failed to connect to github.com port 443: Connection timed
> out

In the rare cases where this happens, it _should_ be transient. After
all, this `curl` command runs after the test suite runs, which happens
after the repository was cloned from github.com. So, if there's a
problem connecting to github.com, it's likely to affect other parts of
the build as well, not just this `curl` command.

This change also adds the `set -x` option so that when a failure occurs,
it will be more clear exactly which command failed.

If all retries fail, the output of the `buildpulse/upload` step will
look something like this:

> + curl -fsSL --retry 3 --connect-timeout 5 https://github.com/buildpulse/test-reporter/releases/latest/download/test-reporter-linux-amd64 > ./buildpulse-test-reporter
> curl: (28) Connection timed out after 5000 milliseconds
>
> Exited with code exit status 28
> CircleCI received exit code 28