Teach orb to gracefully handle failure downloading test-reporter
Prior to this change, a failure to download the test-reporter would
cause the whole build to fail, which is a rotten experience.

We should always be able to successfully download the test-reporter
binary. But we *never* want BuildPulse to make anyone's builds unstable.
So in the (hopefully rare) scenario where we're unable to download the
BuildPulse test-reporter, let's skip the BuildPulse analysis for the
build and emit a warning in the build logs.

Also, in the (again hopefully rare) scenario where we're unable to
download the binary from get.buildpulse.io, let's use github.com as a
backup source for the binary.