Use conditional instead of for loop
ac20d42 used a `for` loop as an attempt to iterate through a set of URLs
for downloading the test-reporter if we fail when downloading from the
first URL. But using a `for` loop this way means that we always fetch
from all URLs. 🤦‍♂️

Instead, we want to use the first test-reporter binary that we
successfully download. So, using conditional instead of a for loop will
give us that behavior.